scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
pandas==2.0.0
APScheduler==3.10.0
lxml==4.9.0
aiohttp==3.8.0
crawl4ai>=0.3.0
